Restructure repository to have multiple examples
We had just thrown this up with different examples in different
branches. This change sets the pattern for how the exampels will work
going forward. There will be effectively separate projects in the
examples/ dir that each have their own project files and integration
test to act as a demo.